Parsoid fragment support: fix handling of 'nowiki' and 'general' strip markers
I originally misunderstood the difference between 'nowiki' and
'general' strip markers, thinking that 'nowiki' was "literal text"
which needed to be HTML escaped, and 'general' was "HTML".  This is
not correct!  Both strip marker types are included as raw HTML; the
'general' strip marker however is technically "half-parsed HTML" and
(in the legacy parser) is subject to doBlockLevels, Language
Conversion, and other processing (T381709).  Parsoid does not
currently have any support for processing "half-parsed HTML" and to
date all code paths involving "half-parsed HTML" have been deprecated.
For the moment, we'll treat the "half-parsed HTML" as "fully-parsed
HTML" in Parsoid.

This patch doesn't make any changes to legacy parser behavior.

Change-Id: I07bacdb4bbe90728d2faa207c19fb92ad0e4a257