mediawiki.diff: Replace pixel font-size value for relative one
This was a relative size before Idd9164e2c5a327.
Not applying Codex design token yet as this would result in bumping up
the font size in various skins by 1px equivalent.
So staying with a hard-coded value for now as long as we don't have
designers and community involved in other direction.

Bug: T313225
Change-Id: Iba79fa430ffb58983d68bb6f67c57174b5c47ccc