Skins: Scan entire navbar for n-specialpages, not just "navigation"
Follow up to:
I03f818d414

We should scan all sections of sidebar instead of just the "navigation" (default) section.

For example, enwiki would likely add the n-specialpages link to the "interaction" (rendered as "contribute") section, and this patch would recognize a link in that section. Previously, they would have to hack the link away by changing MediaWiki:Specialpages-url to "-".

Bug: T333211
Change-Id: I5bb74aaa0febbb6329a4f771d05f8eeed4e8ca83