parser: Replace preg_replace in Parser::finalizeHeadings
preg_match already matches the chars to remove,
no need to run a regex again.
Use a substring with the length of the match (group 0) and
in addition ltrim to handle whitespaces as happen in the replace regex

Convert the regex to constant as it never changes,
it is possible to use constants in constant values.
Remove the declaration of the matcher array, happens by php

Change-Id: I11ff6c004a57f916d3089f9764571fe250e1d93d