auth: Use POST trxProfiler expectations during return/reauth
On AuthManagerSpecialPage, during return (when the user was
redirected to a remote identity provider and back) and reauth
(when the user needed to reenter their credentials before being
given access to the special page), we need to use a GET request
for what we'd normally prefer to be a POST request (because the
user is being redirected back from the remote provider or the
login page, and POST redirects are unreliable). Potentially
the outcome of the request will be some successful authentication,
with all kinds of DB writes involved.

Bug: T385566
Change-Id: I81d1760040af1e10104765bbe803db9b830de289
(cherry picked from commit 4b9eb2823d9e76c4fe4ea24d3bf4c4ea826f5cc6)