refactor(PageUpdatedEvent): tighten return type-hint to match class field
The type of the $page class field is already restricted to
`ProperPageIdentity`, so the getter will always return such.

Aligning these two is useful to avoid an extra pointless go-around in
the code that calls these getter and needs a `ProperPageIdentity` while
the getter returns such but is annotated with `PageIdentity` only.

Change-Id: I8918ba585b9346d2138ee00d4361cb346becf0e2