Deprecate OutputPage::getNoGallery()
Also switch to using OutputPage::$metadata to store $mNoGallery,
although the expected replacement is OutputPage::getOutputFlag().
This is a slight semantic change, as it is the OR of the NO_GALLERY
flag for all ParserOutputs used to compose the Output, instead of
simply being the value of the last ParserOutput used.  This isn't
expected to be significant in practice.

Improved the documentation of ParserOutputFlags::NO_GALLERY while
we're touching this code.

Bug: T301020
Change-Id: Ibaa31df331974e98254ecfcc4393ec1fda2ebe64