Replace a few misplaces is_object checks
is_object is very rarely what we actually need. In many cases we know
exactly what the object can be, and should check for that specific
class or for null.

A database row is an instance of stdClass. Checking that with
is_object also works but is less correct and would allow stuff we
cannot accept in these places.

Change-Id: I1dc663d7325cabc059ef11c4845b0189208e907f