Streamline StatusFormatter::getPsr3MessageAndContext a little
This code was introduced via Id76f554 and later updated via
Iddd14ef. While it's apparently intentional to *not* use an
`instanceof Message` check as that would include unknown
subclasses, I think we can safely assume that RawMessage and
ApiMessage do have a predictable interface.

Both Iddd14ef as well as the test cases in StatusFormatterTest
seem to agree.

I found this while searching for awkward usages of get_class().

Change-Id: I0555e7a92f94a3672833ff0de86d10786ca212db