block: Introduce AbstractBlock::isIndefinite
As far as I can see, the only way to determine
whether a block has an expiry set is through
comparing getExpiry() with the string 'infinity'.
That feels fairly fragile, as well as really
easy to make an error with.

GrowthExperiments needs to introduce a specific
behavior for only indefinitely blocked users.
To be able to make that detection, the abovementioned
method is being added.

Bug: T351234
Change-Id: I85efcc71fe3a6d374b57abaefda8d1d370b41bd6