SpecialBlock [Codex]: Improve form reset logic and add tests
Add a param to the store's resetForm() to not clear internal refs that
are not tied to a specific form field. We call resetForm() with this
option set after a successful form submission.

Add rigorous tests for this flow to ensure no data is leaked from one
block into another.

Bug: T384822
Change-Id: Id9853e10fdfc55ca2e8287cb4bdf20e13b7229eb