objectcache: Introduce dataRedundancy to SqlBagOStuff
This forces SqlBagOStuff to instead of sharding keys, write to n out of
m servers instead. It also reads from those servers as well and in case of
incosistency, picks the value with the highest exptime.

This is mostly for mainstash and allows us to provide stronger
consistency guarantees while allowing for a section to be depooled and
put to maintenance. It basically implements the logic already used by
NoSQL database systems such as Cassandra (There are two types to solve
conflict, quorum or timestamp, Cassandra is using quorum while we are
using timestamp).

There will be some edge cases that it might still pick the wrong value:
 - if TTL is set to INDEF
 - if the TTL gets shortened for various reasons.
 - If we go with two clusters, value is set, one gets depooled, a new
   value is set, the depooled one gets pooled and the other depooled and
   then read happens.
But all of these are extremely rare edge cases and we should be fine.

This also means if data redundancy is set, locking means all sections
will be locked and removal means all sections must allow the unblock.
Otherwise, the lock will be kept.

Bug: T383327
Change-Id: I80da12396858ee4fc58ae257f6c154b3050df696