Fix ContentDOMTransformStageTest after PageBundle refactoring
Since Ibffb2daaf817ce41102211bb9668b29a6e59c0c1, the document creation
passes 'markNew' as a parameter, which allows non-Parsoid output to
round-trip without data-parsoid or ids getting added to them.

This patch adjusts the ContentTransformStageTest to take this into
account, and sets markNew as true explicitly so that we do not depend on
Parsoid vendor patch.

Change-Id: I13bd1316ff22aba5672e500c3b07149d17843811