language: Remove LanguageFactory dependency from MessageCache
Now that MessageParser has been split out, MessageCache has no
remaining calls to Language methods except for getCode().

The only non-deprecated usage of Language is as a parameter to the
callbacks registered by the MessageCacheFetchOverrides hook. Review of
all MessageCacheFetchOverrides handlers in Code Search shows that none
are actually using the Language parameter. None of the closures even
declare a third parameter.

A Language object is a complex and inefficient way to pass a language
code.

So:
* Revert the deprecation of string language codes in the language
  passed to MessageCache::get().
* Pass a string code internally instead of a Language object.
* Relax the validation applied to string language codes, so that it is
  similar to LanguageFactory. Callers were often passing a Language
  object with a code like "qqx" which is not a known language tag, but
  is allowed as a parameter to LanguageFactory::getLanguage().
* Remove the third and fourth parameters to the override callbacks.
* Remove the LanguageFactory constructor parameter.

Depends-On: Ic4beec7373e799e51082957ab96d26e4c692c413
Change-Id: Ia9ddb1b83a6f20d3f7cee2afaf5987dbc99ea00b