language: When using a fallback message, parse with the fallback's language
If you have a message with source like "{{PLURAL:$1|one day|$1 days}}"
and it isn't localised in a language with plural rules that differ from
English, such as Russian, the text "one day" may be displayed to the
user when $1 is say 21, because Russian plural rules are applied to the
English source.

So, keep track of which member of the fallback sequence was used when
fetching a message, and use that source language as the parser target
language when transforming.

In MessageCache::get(), the recently-added $usedKey parameter wasn't
used outside of core, so I repurposed it as a collection of
source-related info.

Bug: T268492
Change-Id: Ic2bfa849d5516dfc55ee7973c6c8097d7e1f8e57