SearchEngine: reset result limit when performing variant searches
Reset the original search limit when performing variant searches, so
that the result set does not get truncated (way) more than necessary.

The `try ... finally` should be unnecessary, but it's generally
advisable to use it to wrap code that messes with the internal state
of a class only to reset it after performing a certain operation.

Add a regression test to verify the fix.

Bug: T386743
Change-Id: I499605b5e0c7b5fc48b919208a1ea3f20fde9d57