user: Allow gender to be a global preference
GenderCache gets the "gender" user preference directly out of the
core database. Thus it does not respect global preferences. So:

* Add UserOptionsLookup::getOptionBatchForUserNames(), which gets a
  single user preference for a set of named users.
* Add UserOptionsStore::fetchBatchForUserNames() as a backend. This is
  a new interface method so is a breaking change.
* Have GenderCache call the new interface.

The no-database case in ServiceWiring is apparently no longer reachable
from the installer due to my previous patch which split
MediaWikiTitleCodec.

Performance generally degrades from one query to three. For change lists
and user page namespace redirects, it shouldn't be a problem, but for
the {{GENDER}} parser function there may be a user-visible performance
degradation.

Bug: T386584
Depends-On: Id02489a597f96cd1cd6db08e16b3624542fdf1f7
Change-Id: I9646b5422afce356e9a1dceeb09d8d4e286dc65e