specials: Add support for file read new in ImageListPager
The queries in this pager can be cleaned up massively on the read new
mode. The same goes for the code which currently does two queries and
then combines the results (and reorders them). Once we can get rid of
the old code, this will allow for removal of hundreds of lines of code.

Bug: T383496
Change-Id: I2427c845b427658788765646df5ad7d4fd7b91ed