Introduce thumbnail steps
This would make it when the editor requests thumbnail with the width of
200px, and we have defined the steps as [...,150px,250px,...], the
thumbnail with the size of 250px to be picked but still the width
attribute in HTML will be set to force the requested size.

This change will massively reduce the storage of thumbnails that have
been causing issues for us recently and improves cache hit ratio in
every layer (from client-side to CDN frontend and backend).

Tested locally and it worked just fine.

Bug: T360589
Change-Id: I9110d4ac9bcd421b07f13deeae5d863ef1ef9c31