CachingSiteStore: Remove needless fallback to getLocalClusterInstance
== Why ==

* Reduce direct use of the internal getLocalClusterInstance,
  so that use cases that actually need it (if any) are found
  and can be addressed in some way.

  https://www.mediawiki.org/wiki/Object_cache#Services

* Make data flow and use of cache service deterministic where.
  It is almost never needed to switch between categorically different
  cache services (eg WAN and MicroStash, or APCU and MainStash) as
  these services already include automatic detection, defaults,
  and fallbacks that should suffice for the use case they serve.

== What ==

This CachingSiteStore service adds caching around DBSiteStore by using
the local server cache (php-apcu, if installed) or a main cache
(wgMainCacheType).

In particular, this service does not (and should not) use CACHE_ANYTHING
or CACHE_DB as fallback; because the thing we are caching is in itself
just a simple database query, so caching it in another database is
counter-productive.

By default, $wgMainCacheType is CACHE_NONE. So out of the box, this
service either does nothing, or, if php-apcu is installed, it uses
that.

In a more complex setup where the sysadmin has installed Memcached
or Redis, as the main cache, that may be used as fallback. Except,
it doesn't because it's very unlikely for a wiki setup to be advanced
enough to include a Memcached or Redis cache, but not even have basic
php-apcu installed.

As such, remove this needless reference to the internal [1] local
cluster cache, and just use getLocalServerObjectCache only.

Note that this class already has an additional fallback in the form
of an in-process cache for repeat calls in the same request, which
is still there after this change.

[1]: https://www.mediawiki.org/wiki/Object_cache#Local_cluster

== Selected history ==

* 2015 (I36b599884): Implement CachingSiteStore, using main cache.
* 2015 (I3eeedddeb): On HHVM, use APCU instead (fix T58602).
* 2016 (I3c25c0ac1): Move construction to ServiceWiring.
* 2018 (I83ff1d2dd): On PHP7, use APCU or main cache (re-fix T58602).

Bug: T186673
Change-Id: I4ae30c17141caae102496b546b919db68a5b93f4