user: Make UserArrayFromResult::$current protected
The property used to be public, but it was made private in
Id90b6838429eca4b. Unfortunately, CentralAuth extends this class and
references the `$current` property directly. So, make it protected and
add a note about CentralAuth being naughty, to prevent this from
happening again.

Follow-up to Id90b6838429eca4b.

Bug: T387148
Change-Id: Ie65381a4fe642ba79f7b845c448e9731dc22647a