Stats: Rewrite BaseMetric for performance
This code is called exceptionally often, sometimes hundreds of
thousands of times. Even small optimizations make a difference.
This patch cuts the relative time spend in this code down in half.

StatsUtils::normalizeString is called much less often.

Slow in_array and array_diff are entirely gone.

The duplicate data structure that was holding half of the data a
second time is gone.

Bug: T378764
Change-Id: I25132ac8db90efc3b4dc5a9bd3c386b9ab4ec327