installer: Migrate web upgrade to the task framework
Web upgrade was the sole remaining caller of restoreServices(), a method
which I accidentally broke. You can't just run RestoredServicesProvider
and expect that to work, it has dependencies.

The order of operations was also backwards. You need to load extensions
before you can reset services, because you need virtual domains from the
extension.json files.

So:

* Port web upgrade to the task framework.
* Move the web upgrade part of DatabaseUpdater::loadExtensions() to
  WebUpgradeExtensionsProvider.
* Move the DatabaseUpdater::doUpdates() call from
  WebInstaller::doUpgrade() to WebUpgradeTask.
* Have WebInstaller::doUpgrade() run the two new tasks with
  RestoredServicesProvider in the middle.

Only default virtual domains are used, not remote mappings; this appears
to be a pre-existing issue.

Bug: T386661
Change-Id: I324fce53d2460b70428115d2b02a31248cc7226e