notifications: Introduce notification aware of Title and Agent
Most of our notifications are aware of Title (the page notification
refers to) and Agent (user who triggered the notification).
Instead of passing those via $custom array we should introduce
a way to pass them directly along with the notification and leave
$custom array to hold only properties that are not yet defined.

Additionally refactor the SimpleNotification to follow same approach.

Bug: T383985
Change-Id: I2f9074256f336ba89abd5aef3322ab0d6d26240d