Restore PHPDoc comment of PageUpdaterFactory::__construct()
Removed in 60941f701c98190be3de66ff7702003dd6507e30, this comment
provides more information than the type declarations—it states the array
value type of `$softwareTags`, not only the fact that it’s an array
(which is sadly all what type declarations can state). So even if this
causes double maintenance, it’s worth being kept because it helps Phan
(and IDEs) flag issues.

Change-Id: I0f88570454e94743eff2c48da2513db9f3fba741