installer: Allow customisation of $wgServer
* Add a box to the name page allowing $wgServer to be reviewed and
  customised.
* Stay on the default server for LocalSettings.php download. Just use
  the customised URL for the "enter your wiki" link. For the Refresh
  header, use a relative URL, since that is safer.

It is a footgun. Ideally there would be client-side validation by
fetching a URL. For now, using a relative URL for the LocalSettings
download means you can fix LocalSettings by editing it.

Also add a little bit of padding between the textbox and the following
label so that the grouping is more clear. Remove the <br> from the
DBConnect page, otherwise there was excessive whitespace in combination
with the new padding.

Bug: T2061
Bug: T64950
Change-Id: I23951ac40faeae9447a2feb2bf5861f0bb81f664