Fix display of SpecialRestSandbox in RTL wikis
Specia:RestSandbox displays generated information that is always
in English; when presented in RTL wikis, that information is
misaligned, and can be completely mangled, especially when seeing
code and json strings.

This patch forces directionality to always be "RTL" with language
setting of "English" to reflect the real language/dir used.

This is similar to what is done in other API generated documentation
like ApiHelp and other places with the same need.

Change-Id: I71067961275b3e0a3c27ec920004f10d20b0fe09