test(ApiStructureTest): use a fresh ApiMain instance for each test
The ApiBase module retains the hookRunner and hookContainer in class
fields, which in turn store created hook handlers in class fields.
Further, the service container will be destroyed after each test in the
\MediaWikiIntegrationTestCase::mediaWikiTearDown.

However, that means if these hook handler have anywhere a reference to
the service container, for example because they get an Object Factory
injected as it is the case with the TaskTypeHandlerRegistry class, then
that service container will be destroyed on subsequent test-runs.

This is a problem if in a subsequent test-run that ObjectFactory is used
to create something that expects to get a service injected. Hence the
error observed in T380585.

This is fixed in this change by creating a new instance of ApiMain for
running the actual test and thus creating all the services fresh from
the new service container for each test.

At least from my local testing, this does not seem to be noticeably
slower than reusing the same object over and over again.

Bug: T380585
Change-Id: I3af04966b58c9ed3c7653f8c5c399135aa2ff341