ResourceLoader: Add page title to user script syntax error
This follows-up Ie309e761f (53a3c8b417), in which we switched from
calling JsMinPlus/JSParser (which took $fileName as argument, and
throw an exception that already included "on line N in X.js")
to calling Peast.

That patch formatted the error as "on line N" but forgot to include
"in X.js".

Test plan:
* Log in as admin and create "MediaWiki:Common.js" with "foo/;",
  which is invalid syntax. Then browse the main page and check the
  browser console.
* Before: "Parse error: Unexpected: ; on line 1"
* After: "Parse error: Unexpected: ; on line 1 in MediaWiki:Common.js"

While at it, document why this cache key is not purely MD5-based,
but also includes wiki and page title.

Change-Id: I16c3b1dacd02ee26a254809e6a9ac60f72aba4da