enotif: Extract email rendering logic
This is a first step to extract the email rendering logic from
EmailNotification. The end goal is to provide a composer class
that can render watchlist notifications.

This is pure refactor and most of code stayed untouched/only moved
from EmailNotification to RecentChangeMailComposer. The only change
was to register all dependencies in composer.

The next step would be use the newly created RecentChangeMailComposer
in Watchlist Notification Handler.

Bug: T387987
Change-Id: Icccfe9548fc67a3692b488947b2648fe41c5198e