block: Increase page restriction limit in multiblocks mode
User consultations uncovered a need for restricting blocks to a list of
more than 10 pages. With multiblocks, an admin in this situation could
split up the block into several blocks, each with 10 pages, which is a
tedious workaround. Discourage this by increasing the limit to 50.

This was already implemented on the client side.

Bug: T388505
Change-Id: I034b9e2fbf38e715255ca84065ec5a63534f86ad