Pass the RecentChange to Notifications
To evolve Watchlist notifications and provide a unified approach
system has to pass the RecentChange object to EmailNotification.
Which later will pass to the RecentChange notification.

This is required to fulfill Echo needs that requires more information
from RecentChange. At current state, Echo requires properties like
`rc_logid` which is not available in MediaWiki ENotif job.

Furthermore this change will simplify the system, as
 * instead of passing multiple params (and adding more of them with
   time) we can just pass one RecentChange object that contains all
   required information. When passing RecentChange via Job we can
   use the `rc_id` to recreate the RecentChange when job gets processed.
 * later Notifications system will be able to listen on RecentChange
   domain event and trigger notifications via listener pattern which
   will simplify the system even more.

Bug: T388665
Change-Id: I2d17bd3743e590e06298efaac387fc6d64e3609d