REST: add translation disclaimer to REST Sandbox
The REST Sandbox is currently implemented in Swagger UI, which does
not support i18n for its own UI elements (it does happily display
our internationalized OpenAPI spec). Add disclaimer text to the
page if it is being viewed in a language other than English, so that
users are aware this is a known issue (and is being worked on).

Bug: T388399
Change-Id: I265aa1e7dce78715dd4977eeb460f1ca26dedbd9