Client-side date/time formatter library
Add a library providing date/time formatting according to the user's
preferred time zone and date preference.

I tested formatting of an example date with all defined formats in all
languages, and I found that it gives identical output to PHP in about
90% of our ~500 languages.

Resolve some of the outstanding issues by aliasing the problematic date
formats on the client side, so that the user will see the date in
another acceptable format for the same language.

The remaining issues mostly relate to the use of a fallback language to
display weekdays and non-Gregorian month names.

Details:

* Add Language::getJsDateFormats(), which converts existing date formats
  to an options array that can be interpreted by the client.
* In Messages*.php, add $numberingSystem, which is the CLDR numbering
  system ID. I set it for all languages that had overriden
  $digitTransformTable. This is sent to the client in the library's JSON
  config and is used as the default numberingSystem option when
  formatting dates.
* In Messages*.php, add $jsDateFormats, which overrides the
  automatically generated date format options.

Bug: T389161
Change-Id: Ib6bc8ebd4d01317aaf32225c6006ea2dc7a1b39e