authmanager: Use an URL parameter to keep track of returns
Remote authentication is split into several requests:
1. Special:Userlogin (or whichever page) calls AuthManager,
   receives a REDIRECT AuthenticationResponse, and redirects to an
   external URL.
2. The remote identity provider, after successfully authenticating
   the user, redirects to Special:Userlogin/returnto.
3. Special:Userlogin/returnto stashes the GET data in the session
   and redirects to Special:Userlogin.
4. Special:Userlogin detects the stashed data, turns it into a
   simulated POST request, and tries to continue authentication.

Steps 2-3 have distinctive URLs but step 4 isn't, it's based on
the presence of session data with a deterministic key, so it could
end up being handled in a request to the same special page in a
different tab. Use a unique GET parameter like we do for
reauthentication to avoid this.

Bug: T388250
Change-Id: Idc9b606cc5b7aca383a5859fc84f182a6aa73385