block: Suppress exception when reblocking a multi-blocked user
If multiple blocks exist for a given target but the wiki is in the
default/old mode with $wgEnableMultiBlocks=false and
$wgUseCodexSpecialBlock=false, attempting to reblock the user causes an
exception.

A race condition in the database has led to a few such multiblocks
existing in production wikis. This issue will soon be resolved by
migrating to multiblocks, so that the blocks will be allowed and can be
easily manipulated. The same error could occur in the future if
$wgEnableMultiBlocks is set to true and then rolled back.

So, catch the exception and show a regular error in the form instead.

Bug: T388743
Change-Id: I67c6ccc20f3c31ba4a4e5734de1641d626adb7f2