objectcache: Re-number array keys in SqlBagOStuff
Currently, if only one server has responded with the key, we take the
first value of the array but the key here is actually the shard index of
the server and can be something like this [ 5 => $value ]. This results
in an null value being used and causing exceptions down the line.

We can simply renumber it to [ 0 => $value ] by calling array_values()

Tested in mwdebug2001 and this fixes the issue

CI tests didn't catch this because we test with a couple of sections
only so shard index ends up being zero anyway.

Bug: T389169
Change-Id: I91a6eab88f28504ef14c8aa3f39bc923ba13e8a7