Download latest phpunit results cache before parallel tests
The PHPUnit parallel test implementation has support for balancing
buckets by duration when timing information is provided for the
test classes.

Update the `phpunit:prepare-parallel:...` action to download the
latest phpunit results cache file and process it as part of the
test suite bucket / `split_group` creation process. If the file
is empty or cannot be download, proceed as before - simply put
the same number of test classes in each bucket.

Bug: T378797
Bug: T384927
Change-Id: Idd167ef8b1d448c236713b1683f768abb6e12655