SpecialBlock [Codex]: validate block ID when passed in via URL parameter
This ensures the target is set properly, and thus auto-loading the form
will also work (as $this->codeFormData['blockTargetExists'] is set).

Use server-provided blockId in SpecialBlock.vue, and move the loading of
blocks from URL param as well as the confirmCloseWindow logic to an
onMounted() hook.

Remove blockId from Jest fixture; this wasn't even being set by the
server but we were pretending it was. Adjust other tests accordingly.

Bug: T389198
Change-Id: Icf5130dc0a873b991513b4034e51ed006659f262