exception: Remove 'exception-json' logging channel
It provided structured exception data in a custom machine-readable
format. Since MediaWiki 1.26, structured exception data conforming to
the PSR-3 logging standard is also available on the 'exception' log
channel, and nobody uses 'exception-json' anymore. But the fact that
it exists and logs errors by default means that it has to be disabled,
and clutters configs, for example:
https://gerrit.wikimedia.org/g/operations/mediawiki-config/+/c0e5c7f71d70de5749f6770abcbbb7e57032bd29/wmf-config/InitialiseSettings.php#2225
https://gerrit.wikimedia.org/g/mediawiki/vagrant/+/c96e609b6ff6a73a412d3fd0a4291259eee026f6/puppet/modules/role/templates/psr3/settings.php.erb#57

Historical context:
* Iacda90fb401f6a45ed1ac1a991e088: Add "exception-json" channel, 2013.
* Ia7ba355d5925a6268ffa321ffc13cc: Add "error-json" channel, 2015.
* I71499d895582bdea033a2516c902e2: Convert "exception" to PSR-3, 2015.
* I782c96dac1181d12267fa610345e87: WMF disables "exception-json", 2017.
* I7a6e6fa52a47a29ec04411a6c8b05e: Remove "error-json" channel, 2024.

The custom data format has been reused for formatting exceptions in
the REST API, so keep getStructuredExceptionData() and rescue the tests.

Bug: T193472
Change-Id: If730618e5371a467cb76db239916b6bbe42512e3