BlockLog.vue: show non-relative expiry where applicable
Log entries now correctly show the timestamp instead of a relavtive
expiry if the date was entered as a timestamp. We aren't able to do the
same for active blocks without a larger refactor, so for now it will
continue to show computed relative expiries even for those entered as
exact timestamps.

Bug: T389275
Change-Id: Iae28dc9cfd6eb7193f7b5cfc7b34fd9537ae2079