qunit: Move 'testrunner' tests to its own file
The environment checks in QUnit.begin() are important to run always,
even when e.g. running with component=EventLogging as filter.

However, the "testrunner" tests are only useful when changing
the assert methods themselves. There is no need to re-run those,
just as there as no need to run any other MediaWiki core tests,
when testing a non-core component.

By moving these out to their own file, part of the 'test.MediaWiki'
module, these still run by default and when selecting "MediaWiki core"
component, but not when selecting another component. Previously,
if you selected e.g. component "EventLogging" the output contained
a visible "testrunner" result, which was not part of that component
which seems surprising/confusing and certainly not needed.

Bug: T250045
Change-Id: I350adf0f9a39796f7052fd5b900acc0ca1a4d4e8