enotif: Retrieve performer and title from RecentChange
This is the second step to refactor EmailNotification
notifyOnPageChange() method to use single RecentChange object
instead multiple arguments.

Previously we had to feed $editor, $title and some attributes.
Instead we can just pass the entire RecentChange and let
EmailNotification decide which properties to use.

In the future, this will allow EmailNotification to listen to
RecentChange domain events to trigger Notifications.

Bug: T388665
Change-Id: I5882531e91cd84c4683522f06d829149fdfc534e