ResourceLoader: Improve error when including file in non-package module
The way ResourceLoaderModule is currently set up, when packagedFiles
are used to declare files, only a "main" script is initially executed,
and other files are executed with require("./file path"). This is not
the same with using "scripts" to declare your module files. Such
scripts are concatenated together into one large script, and require()
can only load modules in this instant.

When you use require() in such script (outside a package file),
we directly use mw.loader.require().

What this fix does is, if require("./*") is encountered in
mw.loader.require(), it throws a descriptive error early, instead of
later in a more confusing way as `Module "." is not loaded`.

Bug: T386833
Change-Id: I97a572c9fefa41e6b97db1d9fd86706490da5c4f