Allow DataAccess::preprocessWikitext() to accept a PFragment as input
The Parsoid PFragment system is roughly analogous to the strip marker
system in core: it allows embedded non-wikitext content within a
wikitext string.  Accept a PFragment as argument to
DataAccess::preprocessWikitext() and convert it to core-style strip
markers in order to tunnel Parsoid fragments through the legacy
preprocessor, legacy extensions, legacy parser functions, etc.

As Parsoid will not actually pass a PFragment to ::preprocessWikitext()
unless $wgParsoidFragmentInput is true, this should be safe to merge.
Round-trip testing can be done w/ the configuration variable set to
true before this is enabled in production generally.

Bug: T268144
Change-Id: Ie543457d5a2eba2ef1f1f4b7622531582c48c3e4