ResourceLoader: Remove unused 'dir' parameter to load.php
When MediaWiki creates load.php URLs, in OutputPage and via RL\Client,
it sets 'lang', not 'dir'. The direction is automatically derived
from the language, which makes the library easier to use for developers,
and also avoids a potentially confusing scenario in which the two
are different (e.g. lang=he and dir=ltr would not make sense, and
would further risk poisoning LTR content in an RTL/Hebrew-related
cache somewhere).

Bug: T225845
Change-Id: I14905c85bb5e3bb47a681dc99668b92844d22f04