UserLookup.vue: trigger new search when changes are made before mounting
Given the Lookup component a template ref and force a lookup and display
the results when the user types in something before Vue loads.

A TODO is left for making the test assert that the menu is visible.
Related: T390127

Move refs to the top of setup() and document them to be consistent with
the structure of other components as well as the store.

SpecialBlock.php: if the Codex form is submitted, treat as no-submission
since this code path isn't supposed to be possible with a JS-only form.

Bug: T389955
Change-Id: I6b9788370495648f1629ec7c3b768e704d78f157